[FIX] l10n_sa_pos: qr code not seen on ticket
When a ticket was sent by email or printed, the qr code was not
visible. This was because it was not correctly put in the report and was
injected in the DOM throught the mounted of the component.

So we are now setting it in the qweb and prepare the values in
export_for_printing instead of inject it through the component that
display it.
