+++ addons/l10n_sa_pos/static/src/xml/OrderReceipt.xml
@@ -3,7 +3,7 @@
     <t t-inherit="point_of_sale.OrderReceipt" t-inherit-mode="extension" owl="1">
         <xpath expr="//img[hasclass('pos-receipt-logo')]" position="after">
             <t if="receipt.is_gcc_country">
-                <div t-if="receipt.qr_code" id="qrcode" class="pos-receipt-logo"/>
+                <img t-if="receipt.qr_code" id="qrcode" t-att-src="receipt.qr_code" class="pos-receipt-logo"/>
                 <br/>
             </t>
         </xpath>
